Trait simple_tlv::Encodable

pub trait Encodable {
    fn encoded_len(&self) -> Result<Length>;

    fn encode(&self, encoder: &mut Encoder<'_>) -> Result<()>;

    fn encode_to_slice<'a>(&self, buf: &'a mut [u8]) -> Result<&'a [u8]> { ... }
}

Encoding trait.

Required methods

fn encoded_len(&self) -> Result<Length>

Compute the length of this value in bytes when encoded as SIMPLE-TLV

fn encode(&self, encoder: &mut Encoder<'_>) -> Result<()>

Encode this value as SIMPLE-TLV using the provided Encoder.

Loading content...
Provided methods

fn encode_to_slice<'a>(&self, buf: &'a mut [u8]) -> Result<&'a [u8]>

Encode this value to the provided byte slice, returning a sub-slice containing the encoded message.

Loading content...