Type Definition simple_tlv::TaggedSlice

type TaggedSlice<'a> = TaggedValue<Slice<'a>>;

Raw SIMPLE-TLV data object TaggedValue<Slice<'_>>.

Implementations

impl<'a> TaggedSlice<'a>

pub fn from(tag: Tag, slice: &'a [u8]) -> Result<Self>

Create a new tagged slice, checking lengths.

pub fn as_bytes(&self) -> &'a [u8]

Borrow the inner byte slice.

pub fn length(&self) -> Length

Get the length of the inner byte slice.

pub fn is_empty(&self) -> bool

Is the inner byte slice empty?

pub fn decode_nested<F, T>(&self, f: F) -> Result<T> where
F: FnOnce(&mut Decoder<'a>) -> Result<T>,

Decode nested values, creating a new Decoder for the data contained in the sequence's body and passing it to the provided FnOnce.